cqlEditor package moved outside query package
